Add a #success? method to API response values
NOTE: THIS IS A BREAKING CHANGE.  Two important changes have been made.
The first is in the result of a failed login.  Previously it was blank.
Now it is the combined service and API error message, to be consistent
with the remainder of the API.  You can tell the difference between a
login which succeeded and one which failed by calling
session_token.success?.  If #success? returns false, the login token
contains an error message.

The second change is in the error message format for login failures.
Previously it would only contain the API error.  Now it contains both
the API and the service error.  For instance:

Old value: "PRODUCT_REQUIRES_FUNDED_ACCOUNT"
New value: "API_ERROR - PRODUCT_REQUIRES_FUNDED_ACCOUNT"

Old value: ""
New value: "INVALID_USERNAME_OR_PASSWORD - OK"