Fix gem dependency versions
Savon broke its log API in version 0.9.10, so here we depend on a
specific older version of savon and savon_spec in both the Gemfile and
the gemspec.  Note: the specs don't seem to pass in this version. I
haven't identified what changed, but the next patch contains fixes to
get them passing again.