Fix logins in the specs
At some point, the specs have broken.  They were attempting to make
actual HTTP connections for the #login calls.  I've added a helper
method in spec_helper.rb to mock it out and updated the before blocks in
the specs to use it.  With this, the specs pass for me.